zstdgpu_tests: widen libzstdgpu heap-size out-params to uint64_t
The underlying zstdgpu_GetGpuMemoryRequirement API takes uint64_t* for the default/upload/readback heap sizes, but the libzstdgpu.h wrapper (both the C API and the C++ TryGetHeapMemorySizes helper) was still declaring these as uint32_t*. That mismatch broke the build with C2664 at libzstdgpu.cpp:146.

Widen the whole chain to uint64_t*:

- libzstdgpu.h: C API decl and TryGetHeapMemorySizes helper

- libzstdgpu.cpp: GetHeapMemorySizesForStage impl

- gpuwork.cpp: DecompressionStageResources fields

shaderVisibleDescriptorCount stays uint32_t (it's a count, not a byte size). CreateHeap already takes size_t, so no downstream changes needed beyond the local variables that receive the sizes.

+++ zstd/zstdgpu_tests/gpuwork.cpp
@@ -103,11 +103,11 @@ Frames ZstdDecompressionWork::Decompress(std::vector<uint8_t>& compressedData, O
     struct DecompressionStageResources
     {
         com_ptr<ID3D12Heap> defaultHeap;
-        uint32_t defaultHeapSizeBytes = 0;
+        uint64_t defaultHeapSizeBytes = 0;
         com_ptr<ID3D12Heap> uploadHeap;
-        uint32_t uploadHeapSizeBytes = 0;
+        uint64_t uploadHeapSizeBytes = 0;
         com_ptr<ID3D12Heap> readbackHeap;
-        uint32_t readbackHeapSizeBytes = 0;
+        uint64_t readbackHeapSizeBytes = 0;
         com_ptr<ID3D12DescriptorHeap> shaderVisibleDescriptorHeap;
         uint32_t shaderVisibleDescriptorCount = 0;
     };
